requirements.txt
wand
Pillow
nltk
pytesseract
PyPDF2
